+++ client/next.config.ts
@@ -1,14 +1,6 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
-	async rewrites() {
-		return [
-			{
-				source: '/api/:path*',
-				destination: `${process.env.API_URL}/:path*`,
-			},
-		];
-	},
 	allowedDevOrigins: ['local.domain'],
 	images: {
 		remotePatterns: [