+++ client/src/api/auth/session.api.ts
@@ -2,14 +2,14 @@ import { ISession } from '@/types';
 import { fetcher } from '@/utils';
 
 export const getAllSessions = async () =>
-	await fetcher<ISession[]>(`/api/v1/auth/sessions/all`);
+	await fetcher<ISession[]>(`/v1/auth/sessions/all`);
 
 export const terminateSession = async (id: string) =>
-	await fetcher<boolean>(`/api/v1/auth/sessions/session/${id}`, {
+	await fetcher<boolean>(`/v1/auth/sessions/session/${id}`, {
 		method: 'DELETE',
 	});
 
 export const terminateAllSessions = async (excludeId: string) =>
-	await fetcher<boolean>(`/api/v1/auth/sessions/except/${excludeId}`, {
+	await fetcher<boolean>(`/v1/auth/sessions/except/${excludeId}`, {
 		method: 'DELETE',
 	});