+++ client/src/api/inventory/inventory.api.ts
@@ -4,10 +4,10 @@ import { fetcher } from '@/utils';
 
 export const getInventory = async (query: PaginationWithLanguageDto) =>
 	await fetcher<IPaginatedDataResponse<IInventory>>(
-		`/api/v1/inventory/me?${new URLSearchParams(Object.entries(query).map(([key, value]) => [key, String(value)]))}`,
+		`/v1/inventory/me?${new URLSearchParams(Object.entries(query).map(([key, value]) => [key, String(value)]))}`,
 	);
 
 export const equipInventoryItem = async (itemId: string) =>
-	await fetcher<IInventory>(`/api/v1/inventory/${itemId}/equip`, {
+	await fetcher<IInventory>(`/v1/inventory/${itemId}/equip`, {
 		method: 'PATCH',
 	});