+++ client/src/api/promotion/promotion.api.ts
@@ -2,4 +2,4 @@ import { UsePromotionDto } from '@/dto';
 import { fetcher } from '@/utils';
 
 export const apiUsePromotion = async ({ alias }: UsePromotionDto) =>
-	await fetcher(`/api/v1/promotion-usage/${alias}`);
+	await fetcher(`/v1/promotion-usage/${alias}`);