+++ client/src/api/report/report.api.ts
@@ -2,7 +2,7 @@ import { SendReportDto } from '@/dto';
 import { fetcher } from '@/utils';
 
 export const sendReport = async (dto: SendReportDto) =>
-	await fetcher('/api/v1/reports/send', {
+	await fetcher('/v1/reports/send', {
 		method: 'POST',
 		body: JSON.stringify({ ...dto }),
 	});